fix(*): ensure duplicate lines are removed without issues
This commit fixes a bug where lines of different lengths were being
erroneously concatenated together. It also fixes a bug where trailing
lines are preserved. Lastly, it leverages the `OrderedDict` class to
make the operations more Pythonic, which makes it a major refactor that
changes the logic of the entire plugin.

fixes #15
fixes #14
potentially fixes #13